feat(mock-doc): get attributes by index
Closes #1211
